move additional test utilties
