+++ src/test/java/org/codeforamerica/shiba/application/parsers/ApplicationDataParserTest.java
@@ -1,7 +1,7 @@
 package org.codeforamerica.shiba.application.parsers;
 
 import org.codeforamerica.shiba.pages.data.ApplicationData;
-import org.codeforamerica.shiba.pages.data.TestApplicationDataBuilder;
+import org.codeforamerica.shiba.testutilities.TestApplicationDataBuilder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
