+++ src/test/java/org/codeforamerica/shiba/output/caf/SnapExpeditedEligibilityDeciderTest.java
@@ -3,7 +3,7 @@ package org.codeforamerica.shiba.output.caf;
 import org.codeforamerica.shiba.Money;
 import org.codeforamerica.shiba.application.parsers.GrossMonthlyIncomeParser;
 import org.codeforamerica.shiba.pages.data.ApplicationData;
-import org.codeforamerica.shiba.pages.data.TestApplicationDataBuilder;
+import org.codeforamerica.shiba.testutilities.TestApplicationDataBuilder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
