+++ src/test/java/org/codeforamerica/shiba/pages/PageControllerTest.java
@@ -1,7 +1,7 @@
 package org.codeforamerica.shiba.pages;
 
 import org.codeforamerica.shiba.ApplicationStatusUpdater;
-import org.codeforamerica.shiba.NonSessionScopedApplicationData;
+import org.codeforamerica.shiba.testutilities.NonSessionScopedApplicationData;
 import org.codeforamerica.shiba.application.Application;
 import org.codeforamerica.shiba.application.ApplicationFactory;
 import org.codeforamerica.shiba.application.ApplicationRepository;
