+++ src/test/java/org/codeforamerica/shiba/pages/enrichment/DateOfBirthEnrichmentTest.java
@@ -2,7 +2,7 @@ package org.codeforamerica.shiba.pages.enrichment;
 
 import org.codeforamerica.shiba.pages.data.ApplicationData;
 import org.codeforamerica.shiba.pages.data.PagesData;
-import org.codeforamerica.shiba.pages.data.TestApplicationDataBuilder;
+import org.codeforamerica.shiba.testutilities.TestApplicationDataBuilder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
