+++ src/test/java/org/codeforamerica/shiba/testutilities/AbstractRepositoryTest.java
@@ -1,6 +1,5 @@
 package org.codeforamerica.shiba.testutilities;
 
-import org.codeforamerica.shiba.NonSessionScopedApplicationData;
 import org.codeforamerica.shiba.application.ApplicationDataEncryptor;
 import org.codeforamerica.shiba.application.ApplicationRepository;
 import org.codeforamerica.shiba.application.StringEncryptor;
